+++ src/proto_016_PtMumbai/lib_plugin/plugin_registerer.ml
@@ -23,9 +23,9 @@
 (*                                                                           *)
 (*****************************************************************************)
 
-module Filter = struct
-  module Proto = Registerer.Registered
-  module Mempool = Plugin.Mempool
+module Validation = struct
+  include Registerer.Registered
+  module Plugin = Plugin.Mempool
 end
 
 module RPC = struct
@@ -39,8 +39,8 @@ module Metrics = struct
   let hash = Registerer.Registered.hash
 end
 
-let () = Shell_plugin.register_filter (module Filter)
+let () = Protocol_plugin.register_validation_plugin (module Validation)
 
-let () = Shell_plugin.register_rpc (module RPC)
+let () = Protocol_plugin.register_rpc (module RPC)
 
-let () = Shell_plugin.register_metrics (module Metrics)
+let () = Protocol_plugin.register_metrics (module Metrics)